Do not run all filters on post titles.
This causes problems with HTML being inserted in ATOM headers (for instance if you have an ACRONYM, it is wrapped in a span with a class of "caps").

Since there is no easy way to remove this potentially destructive behavior, we're going to comment it out. Your theme can always re-implement it on a case-by-case basis with an easy plugin hook.